Gamertag/console logic; shorten split
If we don't know about a gamertag and a console is passed,
make sure the gamertag isn't still given in the arguments by
checking each argument for a gamertag and if there is one,
make that the nick and take it out of the stats. (If a gamertag
is still passed after the console, treat it as normal)

Also, split only returns k/d by default now, to reduce ridiculous
output size.